base analyta
