+++ tests/fixtures/base_analyte/samples/ACC8074A1.xml
