+++ tests/fixtures/base_analyte/samples/ACC8074A2.xml
