+++ tests/fixtures/base_analyte/samples/ACC8075A1.xml
