+++ tests/fixtures/base_analyte/samples/ACC8083A1.xml
